+++ include/Ark/VM/VM.hpp
@@ -54,10 +54,6 @@ namespace Ark
          */
         explicit VM(State& state) noexcept;
 
-        [[deprecated("Use VM(State&) instead of VM(State*)")]] explicit VM(State* state) noexcept :
-            VM(*state)
-        {}
-
         /**
          * @brief Run the bytecode held in the state
          *